requirements.txt
# List of libraries to install
# Type in terminal: pip install -r requirements.txt

requests
beautifulsoup4
python-dotenv